+++ files/default/tests/minitest/create_test.rb
@@ -2,7 +2,6 @@ require File.expand_path('../support/helpers', __FILE__)
 require 'f5-icontrol'
 
 describe_recipe 'f5-bigip::create' do
-
   include Helpers::F5Icontrol
 
   client = nil
@@ -31,7 +30,7 @@ describe_recipe 'f5-bigip::create' do
   describe 'f5_ltm_node' do
     it 'creates a node' do
       node = '10.10.10.13'
-      nodes = client['LocalLB.NodeAddressV2'].get_list.find { |n| n =~ /(^|\/)#{node}$/ }
+      nodes = client['LocalLB.NodeAddressV2'].get_list.find { |n| n =~ %r{(^|\/)#{node}$} }
 
       refute_empty nodes, "Node #{node} not created"
 
@@ -83,7 +82,7 @@ describe_recipe 'f5-bigip::create' do
       refute_kind_of(String, monitors, "Expected pool 'new' health checks to be https and udp")
 
       result = (monitors.size == 2 &&
-               (monitors.uniq.sort == ['/Common/https', '/Common/udp'] || monitors.uniq.sort == %w('https' 'udp')))
+               (monitors.uniq.sort == %w(/Common/https /Common/udp) || monitors.uniq.sort == %w(https udp)))
       assert result,
              "Expected pool 'new' health checks to be https and udp"
     end
@@ -102,7 +101,7 @@ describe_recipe 'f5-bigip::create' do
   describe 'f5_ltm_virtual_server' do
     it 'creates a virtual server' do
       vs = 'vs_new'
-      vs_list = client['LocalLB.VirtualServer'].get_list.find { |v| v =~ /(^|\/)#{vs}$/ }
+      vs_list = client['LocalLB.VirtualServer'].get_list.find { |v| v =~ %r{(^|\/)#{vs}$} }
 
       refute_nil vs_list, "Virtual Server #{vs} not created"
 
@@ -217,9 +216,7 @@ describe_recipe 'f5-bigip::create' do
   describe 'f5_ltm_monitor' do
     it 'creates a new monitor templates' do
       exp_monitors = ['/Common/mon_new', '/Common/mon_new_defaults', '/Common/mon_delete'].sort.uniq
-      monitors = client['LocalLB.Monitor'].get_template_list
-                                          .map { |m| m['template_name'] }
-                                          .sort.uniq
+      monitors = client['LocalLB.Monitor'].get_template_list.map { |m| m['template_name'] }.sort.uniq
 
       assert array_contains?(monitors, exp_monitors),
              "Expected monitors of #{exp_monitors.inspect} to be included in #{monitors.inspect}"