+++ files/default/tests/minitest/provisioner_test.rb
@@ -2,7 +2,6 @@ require File.expand_path('../support/helpers', __FILE__)
 require 'f5-icontrol'
 
 describe_recipe 'f5-bigip::default' do
-
   include Helpers::F5Icontrol
 
   client = nil
@@ -102,11 +101,8 @@ describe_recipe 'f5-bigip::default' do
       # Check format of monitor string
       monitor_string_check = monitors
       monitor_string_check = monitors.first if monitors.is_a? Array
-      if monitor_string_check.include? '/Common/'
-        monitor_check = '/Common/udp'
-      else
-        monitor_check = 'udp'
-      end
+      monitor_check = 'udp'
+      monitor_check = '/Common/udp' if monitor_string_check.include? '/Common/'
 
       if monitors.is_a? Array
         assert_includes(monitors, monitor_check, "Expected pool 'bad_dns' to include udp")
@@ -272,8 +268,7 @@ describe_recipe 'f5-bigip::default' do
 
     it 'deletes a monitor template' do
       monitors_to_delete = ['mon_delete']
-      monitors = client['LocalLB.Monitor'].get_template_list
-                                          .map! { |m| m['template_name'].gsub('/Common/', '') }
+      monitors = client['LocalLB.Monitor'].get_template_list.map! { |m| m['template_name'].gsub('/Common/', '') }
       monitors_missed = monitors_to_delete & monitors
       assert_empty monitors_missed, "These virtual servers still exist: #{monitors_missed}. Not deleted?"
     end