+++ files/default/tests/minitest/support/pool.rb
@@ -8,12 +8,11 @@ module Helpers
         refute true, "Received exception getting pool for #{vs}, probably does not exist"
       end
 
-      assert_match((/(^|\/)#{exp_pool}$/), curr_pool)
+      assert_match(%r{(^|\/)#{exp_pool}$}, curr_pool)
     end
 
     def check_pool_member(client, pool, exp_member)
-      members = client['LocalLB.Pool'].get_member_v2([pool]).first
-                .map { |m| { 'address' => m['address'], 'port' => m['port'] } }
+      members = client['LocalLB.Pool'].get_member_v2([pool]).first.map { |m| { 'address' => m['address'], 'port' => m['port'] } }
 
       assert members.include?(exp_member),
              "Expected to find pool member #{exp_member.inspect} but found the following instead: #{members.inspect}"
@@ -22,8 +21,7 @@ module Helpers
     end
 
     def check_pool_members(client, pool, exp_members)
-      members = client['LocalLB.Pool'].get_member_v2([pool]).first
-                .map { |m| { 'address' => m['address'], 'port' => m['port'] } }
+      members = client['LocalLB.Pool'].get_member_v2([pool]).first.map { |m| { 'address' => m['address'], 'port' => m['port'] } }
 
       assert arrays_match?(members, exp_members),
              "Expected to find pool member #{exp_members.inspect}, found #{members.inspect} instead"