+++ libraries/resource_ltm_irule.rb
@@ -24,7 +24,7 @@ class Chef
     #
     # Chef Resource for F5 LTM iRule
     #
-    class  F5LtmIrule < Chef::Resource
+    class F5LtmIrule < Chef::Resource
       def initialize(name, run_context = nil)
         super
         @resource_name = :f5_ltm_irule
@@ -36,7 +36,7 @@ class Chef
         @irule_name = name
         @template = nil
         @content = nil
-        @variables = []
+        @variables = {}
       end
 
       def irule_name(arg = nil)
@@ -52,7 +52,7 @@ class Chef
       end
 
       def variables(arg = nil)
-        set_or_return(:variables, arg, :kind_of => [Array,Hash], :required => false)
+        set_or_return(:variables, arg, :kind_of => [Array, Hash], :required => false)
       end
 
       def f5(arg = nil)