+++ CHANGELOG.md
@@ -2,6 +2,14 @@ f5-bigip Cookbook CHANGELOG
 ==============================
 This file is used to list changes made in each version of the f5-bigip cookbook.
 
+v0.5.7
+------
+* Rubocop style cleanup
+
+v0.5.6
+------
+* Bug - irule variables hash defaults to an empty array
+
 v0.5.5
 ------
 * Bug - translate address current value overwritten by translate port value
@@ -10,7 +18,12 @@ v0.5.5
 
 v0.5.3
 ------
-* Additional bits for hmh cookbook support
+* Support removing pool members
+* Set username and password for monitors (sets password on every run, because encryption)
+
+v0.5.2
+------
+* Remove subdirs in libraries folder (#9)
 
 v0.5.1
 ------